More cleanup of event logic
