+++ packages/outline-react/src/useOutlinePlainText.js
@@ -9,6 +9,7 @@
 
 import type {OutlineEditor} from 'outline';
 import type {EventHandlerState} from 'outline-react/OutlineEventHandlers';
+import type {InputEvents} from 'outline-react/useOutlineEditorEvents';
 
 import {useEffect, useMemo} from 'react';
 import {createTextNode} from 'outline';
@@ -17,6 +18,23 @@ import {
   createParagraphNode,
   ParagraphNode,
 } from 'outline-extensions/ParagraphNode';
+import {CAN_USE_BEFORE_INPUT} from 'outline-react/OutlineEnv';
+import {
+  onSelectionChange,
+  onKeyDown,
+  onKeyUp,
+  onPointerDown,
+  onPointerUp,
+  onCompositionStart,
+  onCompositionEnd,
+  onCut,
+  onCopy,
+  onNativeBeforeInput,
+  onPastePolyfill,
+  onDropPolyfill,
+  onDragStartPolyfill,
+  onPolyfilledBeforeInput,
+} from 'outline-react/OutlineEventHandlers';
 
 function initEditor(editor: OutlineEditor): void {
   editor.update((view) => {
@@ -30,11 +48,36 @@ function initEditor(editor: OutlineEditor): void {
   });
 }
 
+const emptyObject: {} = {};
+
+const events: InputEvents = [
+  ['selectionchange', onSelectionChange],
+  ['keydown', onKeyDown],
+  ['keyup', onKeyUp],
+  ['pointerdown', onPointerDown],
+  ['pointerup', onPointerUp],
+  ['pointercancel', onPointerUp],
+  ['compositionstart', onCompositionStart],
+  ['compositionend', onCompositionEnd],
+  ['cut', onCut],
+  ['copy', onCopy],
+];
+
+if (CAN_USE_BEFORE_INPUT) {
+  events.push(['beforeinput', onNativeBeforeInput]);
+} else {
+  events.push(
+    ['paste', onPastePolyfill],
+    ['drop', onDropPolyfill],
+    ['dragstart', onDragStartPolyfill],
+  );
+}
+
 export default function useOutlinePlainText(
   editor: OutlineEditor,
   isReadOnly: boolean = false,
 ): {} | {onBeforeInput: (SyntheticInputEvent<EventTarget>) => void} {
-  const pluginState: EventHandlerState = useMemo(
+  const eventHandlerState: EventHandlerState = useMemo(
     () => ({
       isReadOnly: false,
       richText: false,
@@ -45,8 +88,8 @@ export default function useOutlinePlainText(
   );
 
   useEffect(() => {
-    pluginState.isReadOnly = isReadOnly;
-  }, [isReadOnly, pluginState]);
+    eventHandlerState.isReadOnly = isReadOnly;
+  }, [isReadOnly, eventHandlerState]);
 
   useEffect(() => {
     if (editor !== null) {
@@ -55,5 +98,13 @@ export default function useOutlinePlainText(
     }
   }, [editor]);
 
-  return useOutlineEditorEvents(editor, pluginState);
+  useOutlineEditorEvents(events, editor, eventHandlerState);
+
+  return CAN_USE_BEFORE_INPUT
+    ? emptyObject
+    : {
+        onBeforeInput: (event: SyntheticInputEvent<EventTarget>) => {
+          onPolyfilledBeforeInput(event, editor, eventHandlerState);
+        },
+      };
 }
